+++ packages/docusaurus-mdx-loader/src/remark/transformLinks/__tests__/index.test.js
@@ -10,15 +10,15 @@ import remark from 'remark';
 import mdx from 'remark-mdx';
 import vfile from 'to-vfile';
 import plugin from '..';
-import slug from '../../slug';
+import transformImage from '../../transformImage';
 
 const processFixture = async (name, options) => {
   const path = join(__dirname, 'fixtures', `${name}.md`);
   const staticDir = join(__dirname, 'fixtures', 'static');
   const file = await vfile.read(path);
   const result = await remark()
-    .use(slug)
     .use(mdx)
+    .use(transformImage, {...options, filePath: path, staticDir})
     .use(plugin, {...options, filePath: path, staticDir})
     .process(file);
